moved in-app notifications to new architecture
