+++ service/src/services/builders/notification.builder.ts
@@ -1,0 +1,5 @@
+import { BaseEventPayload } from '@alkemio/notifications-lib';
+
+export interface NotificationBuilder {
+  buildAndSend(event: BaseEventPayload): Promise<void>;
+}
