+++ service/src/services/builders/role.config.ts
@@ -1,0 +1,12 @@
+import { InAppNotificationCategory } from '@alkemio/notifications-lib';
+import { AuthorizationCredential } from '@alkemio/client-lib/dist/generated/graphql';
+import { UserPreferenceType } from '@alkemio/client-lib';
+
+export type RoleConfig = {
+  category: InAppNotificationCategory;
+  credential: {
+    type: AuthorizationCredential;
+    resourceID?: string;
+  };
+  preferenceType?: UserPreferenceType;
+};
