+++ service/src/services/builders/utils/payload.builder.fn.ts
@@ -1,0 +1,14 @@
+import {
+  CompressedInAppNotificationPayload,
+  InAppNotificationCategory,
+  InAppNotificationPayloadBase,
+} from '@alkemio/notifications-lib';
+
+export type PayloadBuilderFn<
+  TEvent,
+  TPayload extends InAppNotificationPayloadBase
+> = (
+  category: InAppNotificationCategory,
+  receiverIDs: string[],
+  event: TEvent
+) => CompressedInAppNotificationPayload<TPayload>;
