+++ service/src/services/dispatchers/dispatcher.ts
@@ -1,0 +1,3 @@
+export interface Dispatcher {
+  dispatch(data: any): void;
+}
