+++ service/src/services/dispatchers/in-app/index.ts
@@ -1,0 +1,1 @@
+export * from './in.app.dispatcher';
