+++ service/src/services/dispatchers/index.ts
@@ -1,0 +1,2 @@
+export * from './dispatcher';
+export * from './in-app';
