--- service/src/services/domain/in-app/index.ts
@@ -1,2 +1,0 @@
-export * from './in.app.notification.builder';
-export * from './in.app.notification.builder.module';
