+++ service/src/services/event-subjects/event.subject.ts
@@ -1,0 +1,5 @@
+import { BaseEventPayload } from '@alkemio/notifications-lib';
+
+export interface EventSubject {
+  notifyAll(event: BaseEventPayload): void;
+}
