--- service/src/services/external/in-app-notifications/in.app.notification.sender.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { InAppNotificationSender } from './in.app.notification.sender.service';
-
-@Module({
-  imports: [],
-  providers: [InAppNotificationSender],
-  exports: [InAppNotificationSender],
-})
-export class InAppNotificationSenderModule {}
