--- service/src/services/external/in-app-notifications/index.ts
@@ -1,2 +1,0 @@
-export * from './in.app.notification.sender.module';
-export * from './in.app.notification.sender.service';
